+++ include/master_element/MasterElement.h
@@ -351,101 +351,6 @@ private:
   ) const;
 };
 
-// 3D Quad 27 subcontrol volume
-class Quad92DSCV : public QuadrilateralP2Element
-{
-public:
-  Quad92DSCV();
-  virtual ~Quad92DSCV() {}
-
-  const int * ipNodeMap(int ordinal = 0);
-
-  void determinant(
-    const int nelem,
-    const double *coords,
-    double *areav,
-    double * error );
-
-private:
-  void set_interior_info();
-
-  double jacobian_determinant(
-    const double *POINTER_RESTRICT elemNodalCoords,
-    const double *POINTER_RESTRICT shapeDerivs ) const;
-
-  std::vector<double> ipWeight_;
-};
-
-// 3D Hex 27 subcontrol surface
-class Quad92DSCS : public QuadrilateralP2Element
-{
-public:
-  Quad92DSCS();
-  virtual ~Quad92DSCS() {}
-
-  void determinant(
-    const int nelem,
-    const double *coords,
-    double *areav,
-    double * error );
-
-  void grad_op(
-    const int nelem,
-    const double *coords,
-    double *gradop,
-    double *deriv,
-    double *det_j,
-    double * error );
-
-  void shifted_grad_op(
-    const int nelem,
-    const double *coords,
-    double *gradop,
-    double *deriv,
-    double *det_j,
-    double * error );
-
-  void face_grad_op(
-    const int nelem,
-    const int face_ordinal,
-    const double *coords,
-    double *gradop,
-    double *det_j,
-    double * error );
-
-  void gij(
-    const double *coords,
-    double *gupperij,
-    double *glowerij,
-    double *deriv);
-
-  const int * adjacentNodes();
-
-  const int * ipNodeMap(int ordinal = 0);
-
-  int opposingNodes(
-    const int ordinal, const int node);
-
-  int opposingFace(
-    const int ordinal, const int node);
-
-  const int* side_node_ordinals(int sideOrdinal) final;
-
-
-private:
-  void set_interior_info();
-  void set_boundary_info();
-
-  template <Jacobian::Direction direction> void
-  area_vector(
-    const double *POINTER_RESTRICT elemNodalCoords,
-    double *POINTER_RESTRICT shapeDeriv,
-    double *POINTER_RESTRICT areaVector ) const;
-
-  std::vector<ContourData> ipInfo_;
-  int ipsPerFace_;
-};
-
 // 2D Tri 3 subcontrol volume
 class Tri2DSCV : public MasterElement
 {
